fix: unify Android and OEM font fallback coverage
