+++ RELEASE_NOTES_v2.3.11.md
@@ -1,0 +1,21 @@
+# 洛书 v2.3.11 测试候选
+
+本版保持 v2.3.8/v2.3.9 正式版 App UI 不变，重点修复同一系统中部分区域仍使用默认字体的问题。
+
+## 统一字体槽位
+
+- 不再把带 `variant="elegant|compact"` 的普通 Android UI 字体族误当成不可替换 fallback。
+- `zh-Hans`、`zh-Hant` 等中文 fallback 在所选字体具备中文覆盖时参与逐槽位替换。
+- 英文 fallback 仅在源字体具备完整英文和数字覆盖时替换。
+- 日文、韩文、阿拉伯文、印度文字、Emoji、符号、图标、数学与音乐字体继续使用系统 fallback，避免方框乱码。
+- 保留目标 ROM 字体 family 和 PostScript 内部名称，兼容直接按字体内部名称加载的系统组件与第三方应用。
+- 不同系统 family 会生成独立字体文件，不再因为轮廓参数相同而错误共用一个内部名称。
+- 继续保留 v2.3.10 的挂载验证误报修复：路径、inode 或字体服务视图不同不再判定字体失败。
+
+## 版本信息
+
+- 模块版本：`v2.3.11`
+- 模块 versionCode：`20311`
+- App versionName：`2.3.11`
+- App versionCode：`2031101`
+- 当前类型：测试候选
