+++ config/version_notes.conf
@@ -1,3 +1,3 @@
-version=v2.3.10
-summary=紧急修复：字体已经生效却被误判失败并恢复默认字体
-notes=v2.3.10 保持 v2.3.8/v2.3.9 正式版 App UI 不变。验证器优先读取私有 .luoshu-payload，不再把公开兼容目录的空视图或旧视图当作真实字体源；系统可见字体文件完全一致仍作为强证据，但 KernelSU、metamodule 或 Android 字体服务造成的路径、inode 与文件布局差异不再直接判定失败；PID 1 挂载可见或挂载事务已提交且私有字体负载存在时确认字体已应用；只有自挂载明确失败且没有更强可见证据时才显示失败，避免已生效字体被标红、连续计数并错误恢复为系统默认字体。模块 versionCode 20310，App versionCode 2031001。
+version=v2.3.11
+summary=统一同一系统内的中文、英文和 OEM UI 字体槽位，修复部分区域仍使用默认字体
+notes=v2.3.11 保持 v2.3.8/v2.3.9 正式版 App UI 不变。参考成熟字体模块的“字体文件 + fontconfig + 系统 fallback”方案，不再把所有带 lang/variant 的 Android 字体族一律排除：variant=elegant/compact 的普通 UI 字体会正常替换；zh-Hans/zh-Hant 等中文 fallback 在源字体具备中文覆盖时纳入逐槽位生成；日文、韩文、阿拉伯文、Emoji、符号和源字体不支持的脚本继续使用系统 fallback。生成字体保留目标系统 family/PostScript 内部名称，修复应用按字体内部名称直取时回退默认字体；不同 family 不再错误复用同一个生成文件。包含 v2.3.10 的已生效字体误报修复。模块 versionCode 20311，App versionCode 2031101。
