+++ module.prop
@@ -1,7 +1,7 @@
 id=LuoShu
 name=洛书
-version=v2.3.10
-versionCode=20310
+version=v2.3.11
+versionCode=20311
 author=惜故里丶
-description=Android 无 Hook 全局字体引擎：设备字体清单、私有自挂载、复合字体与安全事务回滚
+description=Android 无 Hook 全局字体引擎：逐系统字体槽位统一替换、私有自挂载与安全回退
 updateJson=https://raw.githubusercontent.com/xgl34222220-ops/LuoShu/main/update.json
