+++ scripts/module_payload_manifest.txt
@@ -13,6 +13,7 @@ common/device_font_load_verify.py
 common/device_font_load_verify.sh
 common/device_font_payload_bridge.sh
 common/device_font_payload_build.py
+common/device_font_payload_build_base.py
 common/device_font_payload_overlay.py
 common/device_font_payload_policy.sh
 common/device_font_payload_runtime.sh
@@ -20,9 +21,12 @@ common/device_font_payload_verify.py
 common/device_font_runtime_report.sh
 common/device_font_transaction_guard.sh
 common/device_font_slot_build.py
+common/device_font_slot_build_base.py
 common/device_font_slot_plan.py
+common/device_font_slot_plan_base.py
 common/device_font_slot_plan.sh
 common/device_font_template.py
+common/device_font_template_base.py
 common/device_font_template.sh
 common/font_archive_export.sh
 common/font_axis_info.py
@@ -41,10 +45,10 @@ common/font_finalize_hotfix.sh
 common/font_import.sh
 common/font_import_compat.sh
 common/font_import_probe.py
+common/font_instance.py
 common/font_inventory.py
 common/font_inventory_scan.py
 common/font_inventory_scan_v3.py
-common/font_instance.py
 common/font_library_cache.sh
 common/font_manager.sh
 common/font_metadata.py
