+++ scripts/v2_source_audit.sh
@@ -56,3 +56,4 @@ grep -q 'native_font_index.json' "$ROOT/service.sh"
 sh "$ROOT/scripts/uninstall_cleanup_test.sh"
 sh "$ROOT/scripts/self_mount_test.sh"
 sh "$ROOT/scripts/font_runtime_policy_test.sh"
+python3 "$ROOT/scripts/fallback_coverage_policy_test.py"
